requirements.txt
datasets>=2.14.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
torch>=2.0.0
tqdm>=4.65.0
openai>=1.0.0
python-dotenv>=1.0.0
nltk>=3.8.1
sentence-transformers>=2.2.2
requests>=2.31.0
tiktoken>=0.7.0